requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.5
django-crispy-forms==1.10.0
django-environ==0.4.5
django-macaddress==1.7.0
django-rest-auth==0.9.5
django-saml2-auth==2.2.1
django-sass-processor==0.8.2
django-xff==1.3.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
netaddr==0.8.0
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
pysaml2==6.4.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
timeago==1.0.15
urllib3==1.26.2
